:lock: add admin user authentication to all admin routes

+++ pages/api/menu/[id].js
@@ -2,7 +2,8 @@ import dbConnect from "../../../util/mongo"
 import Menu from "../../../models/menu"
 
 export default async function handler(req, res) {
-    const { method, query: { id } } = req
+    const { method, query: { id }, cookies } = req
+    const token = cookies.token
     await dbConnect()
 
     if (method === "GET") {
@@ -15,6 +16,9 @@ export default async function handler(req, res) {
         }
     }
     if (method === "PUT") {
+        if (!token || token !== process.env.token) {
+            return res.status(401).json("You are not authenticated!")
+        }
         try {
             const menuItem = await Menu.findByIdAndUpdate(id, req.body, {
                 new: true,
@@ -28,6 +32,9 @@ export default async function handler(req, res) {
     }
 
     if (method === "DELETE") {
+        if (!token || token !== process.env.token) {
+            return res.status(401).json("You are not authenticated!")
+        }
         try {
             await Menu.findByIdAndDelete(id);
             res.status(200).json("The Menu has been deleted!");
